Move `Server` to a separate file.
